% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_gdx.R
\name{extract_all}
\alias{extract_all}
\title{Extract all selected symbols for all scenarios (long format)}
\usage{
extract_all(cfg, reg)
}
\description{
Extract all selected symbols for all scenarios (long format)
}
