% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_load.R
\name{load_config}
\alias{load_config}
\title{Load and validate configuration (defaults + overrides)}
\usage{
load_config(path = "config/project.yml", verbose = TRUE)
}
\description{
Load and validate configuration (defaults + overrides)
}
