% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_export.R
\name{save_artifacts}
\alias{save_artifacts}
\title{Save a named list of data.tables according to config}
\usage{
save_artifacts(objs, cfg, subdir = NULL)
}
\description{
Save a named list of data.tables according to config
}
